update: separate converting functions into standalone package
